+++ src/components/elements/FeatureCard/index.module.scss
@@ -1,0 +1,20 @@
+.card {
+  width: 100%;
+  margin: auto;
+  text-align: center;
+  box-shadow: 0 2px 5px 2px rgba(0, 0, 0, 0.2);
+}
+
+.icon {
+  margin-top: .5em;
+  width: 64px;
+  height: auto;
+  color: #757575;
+}
+
+.button {
+  margin-bottom: 20px;
+  margin: 1em auto;
+  font-size: 16px;
+  padding: 0.5em 1em;
+}
