+++ src/pages/index.tsx
@@ -1,179 +1,51 @@
 import {
-  Backdrop,
   Box,
-  Button,
-  CircularProgress,
+  Grid,
   Typography,
   useTheme,
 } from '@mui/material';
-import { useState } from 'react';
-
-import {
-  InterlaceModal,
-  PartitionModal,
-  RegionCard,
-  TaskAssignModal,
-  TransferModal,
-} from '@/components';
-
-import { useRegions } from '@/contexts/regions';
-import {
-  AssignmentIcon,
-  InterlaceIcon,
-  PartitionIcon,
-  TransferIcon,
-} from '@/icons';
-import { RegionLocation } from '@/models';
+import FeatureCard from '@/components/elements/FeatureCard';
+import Chart from '@/assets/chart.png';
+import Config from '@/assets/config.png';
+import Split from '@/assets/split.png';
+import Trade from '@/assets/trade.png';
 
 const Home = () => {
   const theme = useTheme();
-  const { regions, loading, updateRegionName } = useRegions();
-
-  const [currentRegionIndex, setCurrentRegionIndex] = useState<number>();
-  const [partitionModalOpen, openPartitionModal] = useState(false);
-  const [interlaceModalOpen, openInterlaceModal] = useState(false);
-  const [assignModalOpen, openAssignModal] = useState(false);
-  const [transferModalOpen, openTransferModal] = useState(false);
-
-  const selectedRegion =
-    currentRegionIndex === undefined ? undefined : regions[currentRegionIndex];
-  const regionSelected = selectedRegion !== undefined;
-
-  const management = [
-    {
-      label: 'partition',
-      icon: PartitionIcon,
-      onClick: () => openPartitionModal(true),
-    },
-    {
-      label: 'interlace',
-      icon: InterlaceIcon,
-      onClick: () => openInterlaceModal(true),
-    },
-    {
-      label: 'transfer',
-      icon: TransferIcon,
-      onClick: () => openTransferModal(true),
-    },
-    {
-      label: 'assign',
-      icon: AssignmentIcon,
-      onClick: () => openAssignModal(true),
-    },
-  ];
-
-  const isDisabled = (action: string): boolean => {
-    if (!selectedRegion) return false;
-    // XcRegions can only be transferred.
-    return (
-      action !== 'transfer' &&
-      selectedRegion.location !== RegionLocation.CORETIME_CHAIN
-    );
-  };
 
   return (
-    <Box sx={{ display: 'flex', height: '100%', gap: '1rem' }}>
-      <Box sx={{ maxWidth: '45rem', flexGrow: 1, overflow: 'auto' }}>
-        <Box>
-          <Typography
-            variant='subtitle2'
-            sx={{ color: theme.palette.text.secondary }}
-          >
-            Manage your cores
-          </Typography>
-          <Typography
-            variant='subtitle1'
-            sx={{ color: theme.palette.text.primary }}
-          >
-            Regions Dashboard
-          </Typography>
-        </Box>
-        <Box
-          sx={{
-            display: 'flex',
-            flexDirection: 'column',
-            gap: '2rem',
-            mt: '1rem',
-          }}
+    <Box>
+      <Box>
+        <Typography
+          variant='subtitle2'
+          sx={{ color: theme.palette.text.secondary }}
         >
-          <Backdrop open={loading}>
-            <CircularProgress />
-          </Backdrop>
-          {regions.map((region, index) => (
-            <Box key={index} onClick={() => setCurrentRegionIndex(index)}>
-              <RegionCard
-                regionMetadata={region}
-                active={index === currentRegionIndex}
-                editable
-                updateName={(name) => updateRegionName(index, name)}
-              />
-            </Box>
-          ))}
-        </Box>
-      </Box>
-      <Box
-        sx={{
-          position: 'fixed',
-          right: '10rem',
-          color: theme.palette.text.secondary,
-          background: theme.palette.background.default,
-          minWidth: 280,
-          height: 500,
-          margin: 'auto',
-          padding: '2rem 3rem',
-        }}
-      >
-        <Typography variant='h1'>Manage</Typography>
-        <Box
-          sx={{
-            display: 'flex',
-            flexDirection: 'column',
-            gap: '1.5rem',
-            marginTop: '3rem',
-            alignItems: 'flex-start',
-          }}
+          Explore all the possibilities RegionX offers
+        </Typography>
+        <Typography
+          variant='subtitle1'
+          sx={{ color: theme.palette.text.primary }}
         >
-          {management.map(({ label, icon: Icon, onClick }, index) => (
-            <Button
-              key={index}
-              sx={{
-                color: theme.palette.text.secondary,
-                textTransform: 'capitalize',
-              }}
-              startIcon={<Icon color={theme.palette.text.secondary} />}
-              disabled={isDisabled(label)}
-              onClick={onClick}
-            >
-              {label}
-            </Button>
-          ))}
-        </Box>
+          Home
+        </Typography>
       </Box>
-      {regionSelected && (
-        <>
-          <PartitionModal
-            open={partitionModalOpen}
-            onClose={() => openPartitionModal(false)}
-            regionMetadata={selectedRegion}
-          />
-          <InterlaceModal
-            open={interlaceModalOpen}
-            onClose={() => openInterlaceModal(false)}
-            regionMetadata={selectedRegion}
-          />
-          <TaskAssignModal
-            open={assignModalOpen}
-            onClose={() => openAssignModal(false)}
-            regionMetadata={selectedRegion}
-          />
-          <TransferModal
-            open={transferModalOpen}
-            onClose={() => openTransferModal(false)}
-            regionMetadata={selectedRegion}
-          />
-        </>
-      )}
-    </Box>
+      <Box marginTop="12em">
+        <Grid spacing={8} container>
+          <Grid container item xs={6}>
+            <FeatureCard title={"Manage your regions"} image={Config} buttonText={"Region Dashboard"} href="/regions" />
+          </Grid>
+          <Grid container item xs={6} direction="column">
+            <FeatureCard title={"Track Blockspace Consumption"} image={Chart} buttonText={"Corespace Weigher"} href="https://www.polkadot-weigher.com/" />
+          </Grid>
+          <Grid container item xs={6} direction="column">
+            <FeatureCard title={"Share a Polkadot Core"} image={Split} buttonText={"SOON"} href="/" />
+          </Grid>
+          <Grid container item xs={6} direction="column">
+            <FeatureCard title={"Trade Coretime"} image={Trade} buttonText={"Explore the market"} href="/" />
+          </Grid>
+        </Grid>
+      </Box>
+    </Box >
   );
 };
